+++ source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/__init__.py
@@ -30,7 +30,7 @@ interpretation of the prim paths for different sensor types:
 +---------------------+---------------------------+---------------------------------------------------------------+
 | Frame Transformer   | /World/robot/base         | Leaf exists and is a physics body (Articulation / Rigid Body) |
 +---------------------+---------------------------+---------------------------------------------------------------+
-| IMU                 | /World/robot/base         | Leaf exists and is a physics body (Rigid Body)                |
+| Imu                 | /World/robot/base         | Leaf exists and is a physics body (Rigid Body)                |
 +---------------------+---------------------------+---------------------------------------------------------------+
 
 """